+++ CHANGELOG.rst
@@ -6,6 +6,20 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <http://keepachangelog.com/en/1.0.0/>`_
 and this project adheres to `Semantic Versioning <http://semver.org/spec/v2.0.0.html>`_.
 
+3.0.0
+-----
+
+Added
+.....
+
+- Add overload to editing::substitute to specify which sides are substituted
+
+Changed
+.......
+
+- Add BondDiscretizationOption::RoundToNearest to IO::split to be consistent with IO::read
+
+
 2.0.1
 -----